Update SSRF scenario, webserver and terraform code
