+++ scenarios/ec2_ssrf/assets/ssrf_app/README.md
@@ -5,12 +5,7 @@ This app was adopted from https://github.com/sethsec/Nodejs-SSRF-App.git with lo
 # Nodejs-SSRF-App
 Nodejs application intentionally vulnerable to SSRF
 
-#Operating Systems
-Ubuntu 14.04 TLS
-
-Kali 2.0
-
-#Download and Setup
+## Download and Setup
 
 ```ShellSession
 seth@ubuntu:/opt# sudo git clone https://github.com/sethsec/Nodejs-SSRF-App.git
@@ -31,4 +26,21 @@ seth@ubuntu:/opt/Nodejs-SSRF-App# sudo nodejs ssrf-demo-app.js
 #
 ##################################################
 
+```
+
+## Build and run in a Docker container
+
+```ShellSession
+❯ git clone git@github.com:sethsec/Nodejs-SSRF-App.git
+❯ cd Nodejs-SSRF-App/
+❯ docker build -t "nodejs-ssrf-app" .
+❯ docker run -it -p 8000:8000 nodejs-ssrf-app:latest
+
+##################################################
+#
+#  Server listening for connections on port:8000
+#  Connect to server using the following url:
+#  -- http://[server]:8000/?url=[SSRF URL]
+#
+##################################################
 ```
