--- scenarios/ec2_ssrf/terraform/data_sources.tf
@@ -1,4 +1,0 @@
-#AWS Account Id
-data "aws_caller_identity" "aws-account-id" {
-  
-}
