+++ scenarios/ec2_ssrf/terraform/vpc.tf
@@ -1,61 +1,43 @@
-resource "aws_vpc" "cg-vpc" {
-  cidr_block = "10.10.0.0/16"
+resource "aws_vpc" "vpc" {
+  cidr_block           = "10.10.0.0/16"
   enable_dns_hostnames = true
   tags = {
-      Name = "CloudGoat ${var.cgid} VPC"
-      Stack = "${var.stack-name}"
-      Scenario = "${var.scenario-name}"
+    Name = "CloudGoat ${var.cgid} VPC"
   }
 }
 #Internet Gateway
-resource "aws_internet_gateway" "cg-internet-gateway" {
-  vpc_id = "${aws_vpc.cg-vpc.id}"
+resource "aws_internet_gateway" "internet_gateway" {
+  vpc_id = aws_vpc.vpc.id
   tags = {
-      Name = "CloudGoat ${var.cgid} Internet Gateway"
-      Stack = "${var.stack-name}"
-      Scenario = "${var.scenario-name}"
+    Name = "CloudGoat ${var.cgid} Internet Gateway"
   }
 }
 #Public Subnets
-resource "aws_subnet" "cg-public-subnet-1" {
+resource "aws_subnet" "public_subnet" {
   availability_zone = "${var.region}a"
-  cidr_block = "10.10.10.0/24"
-  vpc_id = "${aws_vpc.cg-vpc.id}"
+  cidr_block        = "10.10.10.0/24"
+  vpc_id            = aws_vpc.vpc.id
+
   tags = {
-      Name = "CloudGoat ${var.cgid} Public Subnet #1"
-      Stack = "${var.stack-name}"
-      Scenario = "${var.scenario-name}"
+    Name = "CloudGoat ${var.cgid} Public Subnet"
   }
 }
-resource "aws_subnet" "cg-public-subnet-2" {
-  availability_zone = "${var.region}b"
-  cidr_block = "10.10.20.0/24"
-  vpc_id = "${aws_vpc.cg-vpc.id}"
-  tags = {
-      Name = "CloudGoat ${var.cgid} Public Subnet #2"
-      Stack = "${var.stack-name}"
-      Scenario = "${var.scenario-name}"
-  }
-}
+
 #Public Subnet Routing Table
-resource "aws_route_table" "cg-public-subnet-route-table" {
+resource "aws_route_table" "public_subnet" {
+  vpc_id = aws_vpc.vpc.id
   route {
-      cidr_block = "0.0.0.0/0"
-      gateway_id = "${aws_internet_gateway.cg-internet-gateway.id}"
+    cidr_block = "0.0.0.0/0"
+    gateway_id = aws_internet_gateway.internet_gateway.id
   }
-  vpc_id = "${aws_vpc.cg-vpc.id}"
+
   tags = {
-      Name = "CloudGoat ${var.cgid} Route Table for Public Subnet"
-      Stack = "${var.stack-name}"
-      Scenario = "${var.scenario-name}"
+    Name = "CloudGoat ${var.cgid} Route Table"
   }
 }
+
 #Public Subnets Routing Associations
-resource "aws_route_table_association" "cg-public-subnet-1-route-association" {
-  subnet_id = "${aws_subnet.cg-public-subnet-1.id}"
-  route_table_id = "${aws_route_table.cg-public-subnet-route-table.id}"
-}
-resource "aws_route_table_association" "cg-public-subnet-2-route-association" {
-  subnet_id = "${aws_subnet.cg-public-subnet-2.id}"
-  route_table_id = "${aws_route_table.cg-public-subnet-route-table.id}"
+resource "aws_route_table_association" "public_subnet" {
+  subnet_id      = aws_subnet.public_subnet.id
+  route_table_id = aws_route_table.public_subnet.id
 }
